+++ src/Data/Patch/Map.hs
@@ -1,5 +1,6 @@
 {-# LANGUAGE DeriveTraversable #-}
 {-# LANGUAGE FlexibleInstances #-}
+{-# LANGUAGE GeneralizedNewtypeDeriving #-}
 {-# LANGUAGE LambdaCase #-}
 {-# LANGUAGE MultiParamTypeClasses #-}
 {-# LANGUAGE StandaloneDeriving #-}
@@ -16,6 +17,7 @@ import Control.Lens
 import Data.Map (Map)
 import qualified Data.Map as Map
 import Data.Maybe
+import Data.Monoid.DecidablyEmpty
 import Data.Semigroup
 
 -- | A set of changes to a 'Map'.  Any element may be inserted/updated or
@@ -24,11 +26,14 @@ import Data.Semigroup
 newtype PatchMap k v = PatchMap { unPatchMap :: Map k (Maybe v) }
   deriving ( Show, Read, Eq, Ord
            , Foldable, Traversable
+           , DecidablyEmpty
            )
 
 -- | 'fmap'ping a 'PatchMap' will alter all of the values it will insert.
 -- Deletions are unaffected.
 deriving instance Functor (PatchMap k)
+-- | The empty 'PatchMap' contains no insertions or deletions
+deriving instance Ord k => Monoid (PatchMap k v)
 
 -- | @a <> b@ will apply the changes of @b@ and then apply the changes of @a@.
 -- If the same key is modified by both patches, the one on the left will take
@@ -37,8 +42,6 @@ instance Ord k => Semigroup (PatchMap k v) where
   PatchMap a <> PatchMap b = PatchMap $ a `mappend` b --TODO: Add a semigroup instance for Map
   -- PatchMap is idempotent, so stimes n is id for every n
   stimes = stimesIdempotentMonoid
-
-makeWrapped ''PatchMap
 
 -- | Apply the insertions or deletions to a given 'Map'.
 instance Ord k => Patch (PatchMap k v) where
@@ -57,11 +60,6 @@ instance TraversableWithIndex k (PatchMap k) where
   itraverse = itraversed . Indexed
   itraversed = _Wrapped .> itraversed <. traversed
 
--- | The empty 'PatchMap' contains no insertions or deletions
-instance Ord k => Monoid (PatchMap k v) where
-  mempty = PatchMap mempty
-  mappend = (<>)
-
 -- | Returns all the new elements that will be added to the 'Map'
 patchMapNewElements :: PatchMap k v -> [v]
 patchMapNewElements (PatchMap p) = catMaybes $ Map.elems p
@@ -69,3 +67,5 @@ patchMapNewElements (PatchMap p) = catMaybes $ Map.elems p
 -- | Returns all the new elements that will be added to the 'Map'
 patchMapNewElementsMap :: PatchMap k v -> Map k v
 patchMapNewElementsMap (PatchMap p) = Map.mapMaybe id p
+
+makeWrapped ''PatchMap