+++ ChangeLog.md
@@ -4,6 +4,8 @@
 
 * Create `PatchMapWithPatchingMove` variant which supports moves with a patch.
 
+* Create `DecidablyEmpty` subclass of `Monoid`.
+
 ## 0.0.2.0
 
 * Consistently provide: